backend (example)/pyproject.toml
[project]
name = "todo-backend"
version = "1.0.0"
description = "FastAPI backend for Todo Application with JWT authentication"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "asyncpg>=0.30.0",
    "python-jose[cryptography]>=3.3.0",
    "PyJWT>=2.9.0",
    "python-multipart>=0.0.17",
    "python-dotenv>=1.0.0",
    # AI Integration
    "openai>=1.0.0",
]

[project.scripts]
dev = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]
